BACKEND/requirements.txt
Authlib==0.14.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
easyocr==1.7.1
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.4
Flask==2.3.3
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.1
ninja==1.11.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
python-bidi==0.4.2
pytweening==1.0.7
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.2
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.8.30
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
waitress==2.1.2
Werkzeug==2.3.7
zipp==3.16.2
